+++ src/Change.cc
@@ -307,7 +307,7 @@ void *Change(void *process)
   int streamID2 = cdoStreamOpenWrite(cdoStreamName(1), cdoFiletype());
   pstreamDefVlist(streamID2, vlistID2);
 
-  int gridsize = vlistGridsizeMax(vlistID2);
+  size_t gridsize = vlistGridsizeMax(vlistID2);
   double *array = (double*) Malloc(gridsize*sizeof(double));
 
   int tsID1 = 0;
