+++ src/Enlargegrid.cc
@@ -36,14 +36,14 @@ void gen_index(int gridID1, int gridID2, int *index)
   int gridtype1 = gridInqType(gridID1);
   int gridtype2 = gridInqType(gridID2);
 
-  int gridsize2 = gridInqSize(gridID2);
+  size_t gridsize2 = gridInqSize(gridID2);
 
   if ( gridtype1 != gridtype2 )
     cdoAbort("Input streams have different grid types!");
 
   if ( index == NULL ) cdoAbort("Internal problem, index not allocated!");
 
-  for ( int i = 0; i < gridsize2; i++ ) index[i] = -1;
+  for ( size_t i = 0; i < gridsize2; i++ ) index[i] = -1;
 
   if ( gridtype1 == GRID_LONLAT || gridtype1 == GRID_GAUSSIAN )
     {
@@ -183,8 +183,8 @@ void *Enlargegrid(void *process)
 
   int gridID1 = vlistGrid(vlistID1, 0);
 
-  int gridsize1 = gridInqSize(gridID1);
-  int gridsize2 = gridInqSize(gridID2);
+  size_t gridsize1 = gridInqSize(gridID1);
+  size_t gridsize2 = gridInqSize(gridID2);
 
   double *array1 = (double*) Malloc(gridsize1*sizeof(double));
   double *array2 = (double*) Malloc(gridsize2*sizeof(double));
@@ -221,13 +221,13 @@ void *Enlargegrid(void *process)
 
 	  double missval1 = vlistInqVarMissval(vlistID1, varID);
 
-	  for ( int i = 0; i < gridsize2; i++ ) array2[i] = missval1;
-	  for ( int i = 0; i < gridsize1; i++ )
+	  for ( size_t i = 0; i < gridsize2; i++ ) array2[i] = missval1;
+	  for ( size_t i = 0; i < gridsize1; i++ )
 	    if ( gindex[i] >= 0 )
 	      array2[gindex[i]] = array1[i];		
 
 	  size_t nmiss2 = 0;
-	  for ( int i = 0; i < gridsize2; i++ )
+	  for ( size_t i = 0; i < gridsize2; i++ )
 	    if ( DBL_IS_EQUAL(array2[i], missval1) ) nmiss2++;
 
 	  pstreamDefRecord(streamID2, varID, levelID);
